+++ packages/orq-rc/src/models/operations/retrievedatapoint.ts
@@ -244,7 +244,7 @@ export type RetrieveDatapoint2DatasetsResponse200Type = ClosedEnum<
 
 export type RetrieveDatapoint2File = {
   /**
-   * The base64 encoded file data, used when passing the file to the model as a string.
+   * The file data as a data URI string in the format 'data:<mime-type>;base64,<base64-encoded-data>'. Example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'
    */
   fileData: string;
   /**
@@ -2481,7 +2481,7 @@ export const RetrieveDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-07-22T10:23:13.475Z",
+    "2025-07-23T08:49:25.849Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -2539,7 +2539,7 @@ export const RetrieveDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-07-22T10:23:13.475Z"))
+  updated: z.date().default(() => new Date("2025-07-23T08:49:25.849Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {