+++ packages/orq-rc/RELEASES.md
@@ -2579,3 +2579,13 @@ Based on:
 - [typescript v3.10.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.14 - packages/orq-rc
+
+## 2025-07-23 08:53:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.586.0 (2.661.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.16 - packages/orq-rc